feat(db-remove-animal): add checking if animal exists before removing it

+++ src/data/usecases/remove-animal/db-remove-animal.ts
@@ -1,3 +1,4 @@
+import { ILoadAnimalByIdRepository } from "../load-animal/db-load-animal-protocols";
 import {
 	IDbRemoveAnimal,
 	IRemoveAnimalByIdRepository,
@@ -5,10 +6,17 @@ import {
 
 export class DbRemoveAnimal implements IDbRemoveAnimal {
 	constructor(
-		private readonly removeAnimalByIdRepository: IRemoveAnimalByIdRepository
+		private readonly removeAnimalByIdRepository: IRemoveAnimalByIdRepository,
+		private readonly loadAnimalByIdRepository: ILoadAnimalByIdRepository
 	) {}
 
-	async remove(id: string): Promise<boolean> {
-		return await this.removeAnimalByIdRepository.removeAnimal(id);
+	async remove(id: string, ownerId: string): Promise<boolean> {
+		const animal = await this.loadAnimalByIdRepository.loadAnimal(id);
+
+		if (animal && animal.ownerId === ownerId) {
+			return await this.removeAnimalByIdRepository.removeAnimal(id);
+		}
+
+		return false;
 	}
 }
